Add a editable check to the pages

+++ app/helpers/admin/cmspages_helper.rb
@@ -2,8 +2,11 @@ module Admin::CmspagesHelper
   
   def cmspage_links_column(cmspage)
     html = link_to('View', admin_cmspage_path(cmspage), :class => 'member_link')
-    html << link_to('Edit', edit_admin_cmspage_path(cmspage), :class => 'member_link')
-    html << link_to('Delete', admin_cmspage_path(cmspage),:method => :delete ,:class => 'member_link',:confirm => 'Are you sure?')
+    if cmspage.editable?
+      html << link_to('Edit', edit_admin_cmspage_path(cmspage), :class => 'member_link')
+      html << link_to('Delete', admin_cmspage_path(cmspage),:method => :delete ,:class => 'member_link',:confirm => 'Are you sure?')
+    end
+    html
   end
   
 end
